requirements.txt
pyperclip
qrcode
Pillow==9.5.0
pywin32
atexit